tutor-agent/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-cloud-aiplatform
setuptools